refactor(dashboard): split the stream-contract vocabulary out of keeper-state
