+++ dashboard/src/keeper-state.ts
@@ -3,6 +3,8 @@ import { formatKeeperVisibleReply } from './keeper-message'
 import { parseTextToChatBlocks } from './lib/chat-blocks'
 import { isInFlightDelivery } from './lib/keeper-delivery'
 import { isRecord, asString, asNumber, asBoolean, toIsoTimestamp } from './components/common/normalize'
+import { asStrictStringArray, withoutUndefined } from './lib/json-coerce'
+import { keeperStreamContract, normalizeStreamContract } from './keeper-stream-contract'
 import {
   nonBlankToolCallId,
   toolEntryIdFromCallId,
@@ -14,8 +16,6 @@ import type {
   KeeperConversationRole,
   KeeperConversationSource,
   KeeperConversationStreamContract,
-  KeeperConversationStreamContractSource,
-  KeeperConversationStreamContractStatus,
   KeeperConversationStreamState,
   KeeperConversationDelivery,
   KeeperDiagnostic,
@@ -247,127 +247,6 @@ function normalizeAttachments(raw: unknown): KeeperConversationAttachment[] | un
   return atts.length > 0 ? atts : undefined
 }
 
-function normalizeStringArray(raw: unknown): string[] | null {
-  if (!Array.isArray(raw)) return null
-  const values = raw.map((value) => asString(value))
-  if (values.some((value) => value === undefined)) return null
-  return values as string[]
-}
-
-function withoutUndefined<const T extends Record<string, unknown>>(record: T): T {
-  const next: Record<string, unknown> = {}
-  for (const [key, value] of Object.entries(record)) {
-    if (value !== undefined) next[key] = value
-  }
-  return next as T
-}
-
-export function keeperStreamContract(
-  source: KeeperConversationStreamContractSource,
-  status: KeeperConversationStreamContractStatus,
-  opts: {
-    eventName?: string | null
-    requestId?: string | null
-    turnRef?: string | null
-    traceEventCount?: number | null
-    lifecycleEvents?: string[] | null
-    reason?: string | null
-  } = {},
-): KeeperConversationStreamContract {
-  return withoutUndefined({
-    source,
-    status,
-    eventName: opts.eventName ?? undefined,
-    requestId: opts.requestId ?? undefined,
-    turnRef: opts.turnRef ?? undefined,
-    traceEventCount: opts.traceEventCount ?? undefined,
-    lifecycleEvents: opts.lifecycleEvents ?? undefined,
-    reason: opts.reason ?? undefined,
-  })
-}
-
-export function keeperClientObservedSseStreamContract(
-  source: KeeperConversationStreamContractSource,
-  status: KeeperConversationStreamContractStatus,
-  opts: {
-    eventName?: string | null
-    requestId?: string | null
-    turnRef?: string | null
-    traceEventCount?: number | null
-    lifecycleEvents?: string[] | null
-    reason?: string | null
-  } = {},
-): KeeperConversationStreamContract {
-  return keeperStreamContract(source, status, opts)
-}
-
-function normalizeStreamContractSource(value: unknown): KeeperConversationStreamContractSource | null {
-  switch (asString(value)?.trim()) {
-    case 'keeper_chat_store':
-      return 'keeper_chat_store'
-    case 'backend_stream_lifecycle':
-      return 'backend_stream_lifecycle'
-    case 'backend_turn_trace':
-      return 'backend_turn_trace'
-    case 'rest_history':
-      return 'rest_history'
-    case 'sse_event':
-      return 'sse_event'
-    case 'client_operation_store':
-      return 'client_operation_store'
-    case 'client_operation_lookup':
-      return 'client_operation_lookup'
-    case 'client_local_send':
-      return 'client_local_send'
-    case 'client_stream_failure':
-      return 'client_stream_failure'
-    default:
-      return null
-  }
-}
-
-function normalizeStreamContractStatus(value: unknown): KeeperConversationStreamContractStatus | null {
-  switch (asString(value)?.trim()) {
-    case 'backend_stream_event':
-      return 'backend_stream_event'
-    case 'backend_terminal_event':
-      return 'backend_terminal_event'
-    case 'backend_lifecycle_replay':
-      return 'backend_lifecycle_replay'
-    case 'backend_trace_join':
-      return 'backend_trace_join'
-    case 'history_without_turn_ref':
-      return 'history_without_turn_ref'
-    case 'history_without_stream_events':
-      return 'history_without_stream_events'
-    case 'client_operation_terminal':
-      return 'client_operation_terminal'
-    case 'client_placeholder':
-      return 'client_placeholder'
-    case 'client_reconciled_history':
-      return 'client_reconciled_history'
-    case 'contract_gap':
-      return 'contract_gap'
-    default:
-      return null
-  }
-}
-
-function normalizeStreamContract(raw: unknown): KeeperConversationStreamContract | null {
-  if (!isRecord(raw)) return null
-  const source = normalizeStreamContractSource(raw.source)
-  const status = normalizeStreamContractStatus(raw.status)
-  if (source === null || status === null) return null
-  return keeperStreamContract(source, status, {
-    eventName: asString(raw.event_name) ?? asString(raw.eventName) ?? null,
-    requestId: asString(raw.request_id) ?? asString(raw.requestId) ?? null,
-    turnRef: asString(raw.turn_ref) ?? asString(raw.turnRef) ?? null,
-    traceEventCount: asNumber(raw.trace_event_count) ?? asNumber(raw.traceEventCount) ?? null,
-    lifecycleEvents: normalizeStringArray(raw.lifecycle_events) ?? normalizeStringArray(raw.lifecycleEvents) ?? null,
-    reason: asString(raw.reason) ?? null,
-  })
-}
-
 function normalizeTableCell(raw: unknown): ChatTableCellValue | null {
   const text = asString(raw)
   if (text !== undefined) return text
@@ -572,7 +451,7 @@ function normalizeBlocks(raw: unknown, role: KeeperConversationRole): ChatBlock[
           : null
       }
       if (t === 'ul') {
-        const items = normalizeStringArray(item.items)
+        const items = asStrictStringArray(item.items)
         return items && items.length > 0 ? { t: 'ul', items } : null
       }
       if (t === 'callout') {
